All the setup stuff

+++ src/main/java/com/fdmgroup/sprintfourtemp/dal/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fdmgroup.sprintfourtemp.model.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
 }
